Log in case lock needs to be awaited
